+++ fixture/virtual-catalog/virtualCatalog.ts
@@ -4,10 +4,10 @@ import {
     AbstractItem,
     ActionHandler,
     Item
-} from "../../dist/lib/v4/catalog/AbstractCatalog";
-import {Adminizer} from "../../dist";
+} from "../../src/lib/v4/catalog/AbstractCatalog";
+import {Adminizer} from "../../src";
 import {v4 as uuid} from "uuid";
-import {NavItem, StorageServices} from "../../dist/lib/v4/catalog/Navigation";
+import {NavItem, StorageServices} from "../../src/lib/v4/catalog/Navigation";
 
 interface TestItem extends Item {
     modelId?: string